remove first boost in cal angle

+++ state_cache.sh
@@ -13,11 +13,19 @@ fi
 echo "Saving state at ${cache_path}"
 
 cache_file() {
+  if [ ! -n "${1}" ];
+  then
+    return;
+  fi
   if [ ! -d `dirname ${cache_path}/${1}` ];
   then
     mkdir -p `dirname ${cache_path}/${1}`
   fi
-  cp ${1} ${cache_path}/${1}
+  
+  if [ -f ${1} ];
+  then
+    cp ${1} ${cache_path}/${1}
+  fi
 }
 
 json_file=`ls -rt *params*.json`
